chore: move components to root
